+++ src/Auxillary/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(Exception)
 add_subdirectory(Matrixhandler)
 add_subdirectory(Mathfunctions)
+add_subdirectory(Stochastics)
 add_subdirectory(Coloroutput)
 add_subdirectory(Misc)
 add_subdirectory(Input_Output)
