--- src/Auxillary/Mathfunctions/system/.clang-format
@@ -1,1 +1,0 @@
-#
