+++ src/Auxillary/Stochastics/Normaldistribution.cpp
@@ -1,7 +1,6 @@
 #include "Normaldistribution.hpp"
 #include "pcg_extras.hpp"
 #include "randutils.hpp"
-#include <bits/stdint-uintn.h>
 #include <cmath>
 #include <cstdint>
 #include <iomanip>
@@ -30,6 +29,7 @@ namespace Aux {
     // randutils::auto_seed_256 seed_source(args);
     randutils::auto_seed_256 seed_source;
     pcg64 rng(seed_source);
+    std::cout << std::endl;
     return rng;
   }
